requirements.txt
# Text extraction
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
pdf2image>=1.16.0
python-docx>=1.0.0

# OCR - LightOnOCR-2 (state-of-the-art, Apache 2.0)
transformers>=4.36.0
torch>=2.0.0
accelerate>=0.25.0
Pillow>=10.0.0

# Audio transcription
openai-whisper>=20231117

# Named Entity Recognition
spacy>=3.7.0
nameparser>=1.1.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.13.0

# Web scraping & downloading
selenium>=4.15.0
webdriver-manager>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
gdown>=5.0.0
tqdm>=4.66.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
loguru>=0.7.0
jellyfish>=1.0.0  # For phonetic matching (soundex, metaphone)

# Development
pytest>=7.4.0
black>=23.12.0
ruff>=0.1.0
